hopefully fix the yellow
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@22421 a1c6a512-1295-4272-9138-f99709370657

+++ apps/gui/skin_engine/wps_debug.c
@@ -606,8 +606,8 @@ void print_debug_info(struct wps_data *data, enum wps_parse_error fail, int line
 void debug_skin_usage(void)
 {
     if (wps_verbose_level > 1)
-        DEBUGF("Skin buffer usage: %ld/%ld\n", skin_buffer_usage(),
-                                skin_buffer_usage() + skin_buffer_freespace());
+        DEBUGF("Skin buffer usage: %lu/%lu\n", (unsigned long)skin_buffer_usage(),
+                                (unsigned long)(skin_buffer_usage() + skin_buffer_freespace()));
 }
